"Changed floor plans tab to be active by default and updated corresponding CSS; fixed bug with video tab being active by default"
Explanation:

The code diff includes changes to both the CSS and PHP files related to the navigation pills for floor plans, video, and map tabs on a single property page.

In the CSS file, a new class ".nav-pills-a.nav-pills" has been added with display: flex, justify-content: space-between, and width: 100% properties. This is to style the navigation pills in a horizontal layout with equal spacing between them.

In the PHP file, the order of the tabs has been changed so that the floor plans tab is now the first tab and is set to be active by default. The video tab is now the third tab and is no longer active by default. Additionally, there was a bug in the previous version where the floor plans tab was not being displayed even when there were floor plans available. This has been fixed by adding a check for empty planos array and only displaying the floor plans tab when there are floor plans available.

Overall, the changes are aimed at improving the user experience by making the floor plans tab the default tab and ensuring that it is displayed when there are floor plans available.